Move ProgramQuery into own file
